ai-agent/requirements.txt
numpy
pandas
scikit-learn
requests
web3
python-dotenv